tests: predictable and simpler tempfile management
Instead of using completely random names, we'll use
predictable ones since we already depend on them for
exit codes and such.  This drops our ability to run
the same test for the same version of Ruby in the
same working tree, but that's an unlikely scenario.

While we're at it, avoid remove tempfiles if a test
failed.  This should make debugging easier.