README: change ordering of concurrency model listing
Revactor was our first take on the "rainbows" branch of Unicorn
and probably less confusing to people that may have tracked that
branch.  ThreadPool was used in the "gossamer" branch of Unicorn
before, but since this project is named Rainbows!, it's less
logical.  I'll freely admit to having some personal bias here
and not being a fan of threads (and especially not pthreads!).