tests: avoid race condition in reopen logs test
Ensure our workers are capable of processing requests before we
can continue to setup signal handlers.  The later "kill -USR1"
signal may fire in our test script before the workers have a
chance to setup the handlers.  So by the time we receive a
positive response, we _know_ signal handlers are ready.