tests: add a TAP producer shell library
This will eventually allow our test output to be consumed by any
TAP[1]-compatible consumers.  Like the rest of our shell scripts,
this only relies on POSIX behavior and has no bashisms.

mktemp(1) is a requirement only if using the verbose output in
this library.

This library aims to be useful outside of Rainbows!

[1]: Test Anything Protocol: http://testanything.org/