tests: generate all dependencies atomically
Using a "+" suffix alone was not enough protection since we use
evil recursive makes and can't share dependency info with parent
makes.  While this could be done more efficiently (even with
recursive make), but it'd be harder to maintain.  So we generate
the dependencies later to and sacrifice efficiency on the
initial run (but rarely/never again).