Allow 'use "model"' as a string as well as symbol
Since const_get works with a string as well as a symbol,
allow that to be used.  It's easier and simpler to just
allow strings as use arguments than to error check and
raise exceptions.

So both of the following should now work:

  Rainbows! do
    use :Revactor
  end

  Rainbows! do
    use "Revactor"
  end

Rainbows! will always use the symbol variant internally,
however, so applications can alway expect env['rainbows.model']
to be a symbol.