tests: allow use of alternative sha1 implementations
sha1sum(1) is only common GNU systems, and it may be installed
as gsha1sum on *BSDs.  We'll also try using the openssl sha1
implementation, too.  And finally, we'll provide our own Ruby
sha1sum.rb implementation as a last resort.

We go to great lengths to avoid our own Ruby version because we
want to avoid putting too much trust in ourselves, our Ruby
skills, and even the Ruby implementations.  This is especially
with regard to our knowledge and correct usage of Ruby 1.9
encoding support.  It would actually be *easier* to only use
sha1sum.rb and call it a day.  We just choose to support
SHA1 implementations provided by third parties if possible.

Performance is not a factor since sha1sum.rb performance is very
close to the C implementations.