gracefully exit workers if reopening logs fails
Permissions for the logs could've been badly set by the master.
So we we'll let the master reopen them and refork children to
get around this problem.  We have to be more careful when
reopening logs because we can reopen them in the middle of
client requests (we have to) whereas Unicorn has the luxury
of _knowing_ it has no active clients when it does the reopen.