Rainbows! 0.94.0 - one eight ate my homework!
This release fixes corrupted large response bodies for Ruby 1.8
users with the WriterThreadSpawn and WriterThreadPool models
introduced in 0.93.0.  This bug did not affect Ruby 1.9 users
nor the users of any older concurrency models.

There is also a strange new Rainbows::Sendfile middleware.  It
is used to negate the effect of Rack::Contrib::Sendfile, if that
makes sense.  See the RDoc or
http://rainbows.rubyforge.org/Rainbows/Sendfile.html for all the
gory details.

Finally, the RDoc for our test suite is on the website:

  http://rainbows.rubyforge.org/Test_Suite.html

I wrote this document back when the project started but
completely forgot to tell RDoc about it.  Personally, this
test suite is one of my favorite parts of the project.