doc: avoid documenting internals on RDoc website
Since we suck at building websites, we just rely on RDoc as a
website builder.  And since Rainbows! is an application server
(and not a programming library), our internal API should be of
little interest to end users.

Anybody interested in Rainbows! (or any other project) internals
should be reading the source.