event_machine: better handling of staggered pipelines
Since TCP sockets stream, HTTP requests do not come in at
well-defined boundaries and it's possible for pipelined requests
to come in in a staggered form.  We need to ensure our
receive_data callback doesn't fire any actions at all while
responding with a deferrable @body.

We still need to be careful about buffering, since EM does not
appear to allow temporarily disabling read events (without
pausing writes), so we shutdown the read end of the socket
if it reaches a maximum header size limit.