Rainbows! 2.0.0 - minority rules!
This release is targeted at the minority of web applications
that deal heavily with uploads.

Thanks to Unicorn 3.x, we now support HTTP keepalive for
requests with bodies as long as the application consumes them.
Unicorn 3.x also allows disabling the rewindability requirement
of "rack.input" (in violation of the Rack 1.x spec).

The global client_body_max_size may also be applied per-endpoint
using the Rainbows::MaxBody middleware described in:

  http://rainbows.rubyforge.org/Rainbows/MaxBody.html